Class third: Authentication

+++ server/src/routes/memories.ts
@@ -1,9 +1,17 @@
 import { FastifyInstance } from 'fastify'
 import { prisma } from '../lib/prisma'
 import { z } from 'zod'
+
 export async function memoriesRoutes(app: FastifyInstance) {
-  app.get('/memories', async () => {
+  app.addHook('preHandler', async (req) => {
+    await req.jwtVerify()
+  })
+
+  app.get('/memories', async (req) => {
     const memories = await prisma.memory.findMany({
+      where: {
+        userId: req.user.sub,
+      },
       orderBy: {
         createdAt: 'asc',
       },
@@ -16,7 +24,7 @@ export async function memoriesRoutes(app: FastifyInstance) {
     }))
   })
 
-  app.get('/memories/:id', async (req) => {
+  app.get('/memories/:id', async (req, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -28,6 +36,9 @@ export async function memoriesRoutes(app: FastifyInstance) {
         id,
       },
     })
+
+    if (!memory.isPublic && memory.id !== req.user.sub)
+      return reply.status(401).send()
 
     return memory
   })
@@ -47,14 +58,14 @@ export async function memoriesRoutes(app: FastifyInstance) {
         coverUrl: '',
         createdAt: '',
         id: '',
-        userId: '',
+        userId: req.user.sub,
       },
     })
 
     return memory
   })
 
-  app.put('/memories/:id', async (req) => {
+  app.put('/memories/:id', async (req, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -68,7 +79,15 @@ export async function memoriesRoutes(app: FastifyInstance) {
 
     const { content, isPublic } = bodySchema.parse(req.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== req.user.sub) return reply.status(401).send()
+
+    memory = await prisma.memory.update({
       where: {
         id,
       },
@@ -81,12 +100,20 @@ export async function memoriesRoutes(app: FastifyInstance) {
     return memory
   })
 
-  app.delete('/memories/:id', async (req) => {
+  app.delete('/memories/:id', async (req, reply) => {
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(req.params)
+
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where: {
+        id,
+      },
+    })
+
+    if (memory.userId !== req.user.sub) return reply.status(401).send()
 
     await prisma.memory.delete({
       where: {
